Reorganize upload interface
