+++ site.config.json
@@ -3,14 +3,14 @@
   "site_icon": "🔬",
   "author": "IMOD",
   "keywords": "SMLM,PALM,STORM,PAINT,SR,SuperRes,Super resolution,Microscopy,Live Cell Imaging,Bioimaging",
-  "description": "ShareLoc--A platform for sharing localization data",
+  "description": "ShareLoc--A platform for sharing single-molecule localization microscopy data",
   "twitter_site": "@ImodLab",
   "twitter_creator": "@weioyang",
   "twitter_image": "",
   "og_url": "https://shareloc.xyz",
   "og_title": "ShareLoc.XYZ",
   "og_image": "",
-  "og_description": "A platform for sharing localization data",
+  "og_description": "A platform for sharing single-molecule localization microscopy data",
   "background_image": "",
   "manifest_url": "/manifest.shareloc.json",
   "explore_button_text": "🚀 Explore",
@@ -32,7 +32,7 @@
   "rdf_root_repo": "imodpasteur/shareLoc.xyz",
   "enable_comment": true,
   "splash_title": "ShareLoc.XYZ",
-  "splash_subtitle": "A platform for sharing localization data",
+  "splash_subtitle": "A platform for sharing single-molecule localization microscopy data",
   "splash_feature_list": [
   ],
   "table_view": {
