+++ src/components/Dataset.vue
@@ -11,13 +11,18 @@
         open
       >
         <div class="p-1">
-          <div class="block">
-            <img
-              src="https://raw.githubusercontent.com/buefy/buefy/dev/static/img/buefy-logo.png"
-              alt="Lightweight UI components for Vue.js based on Bulma"
-            />
-          </div>
-          <b-menu class="is-custom-mobile">
+          <form-json
+            v-if="jsonFields && jsonFields.length > 0"
+            :btnReset="{ value: 'Reset' }"
+            :btnSubmit="{ value: 'OK' }"
+            :camelizePayloadKeys="false"
+            :formFields="jsonFields"
+            :formName="'metadata'"
+            @formSubmitted="formSubmitted"
+            :components="components"
+          >
+          </form-json>
+          <!-- <b-menu class="is-custom-mobile">
             <b-menu-list label="Menu">
               <b-menu-item
                 icon="information-outline"
@@ -57,30 +62,21 @@
             <b-menu-list label="Actions">
               <b-menu-item icon="logout" label="Logout"></b-menu-item>
             </b-menu-list>
-          </b-menu>
+          </b-menu> -->
         </div>
-        <section style="padding: 20px;">
-          <form-json
-            v-if="jsonFields && jsonFields.length > 0"
-            :btnReset="{ value: 'Reset' }"
-            :btnSubmit="{ value: 'OK' }"
-            :camelizePayloadKeys="false"
-            :formFields="jsonFields"
-            :formName="'metadata'"
-            @formSubmitted="formSubmitted"
-            :components="components"
-          >
-          </form-json>
-        </section>
+        
       </b-sidebar>
-
-      <div class="p-1">
-        {{ resourceId }}
+      <div style="width: 100%;">
+        <drop-files-field class="block" :item="fileField"></drop-files-field>
+        <div class="viewer block" id="preview-container"></div>
       </div>
     </section>
   </div>
 </template>
 <script>
+import JSZip from "jszip";
+import yaml from "js-yaml";
+import spdxLicenseList from "spdx-license-list/full";
 import "vue-form-json/dist/vue-form-json.css";
 import formJson from "vue-form-json/dist/vue-form-json.common.js";
 // import Markdown from "@/components/Markdown.vue";
@@ -91,7 +87,7 @@ import DropFilesField from "@/components/dropFilesField.vue";
 
 export default {
   name: "dataset",
-  props: ["resourceId"],
+  props: ["resourceId", "rdf", "files"],
   components: {
     "form-json": formJson,
     // markdown: Markdown,
@@ -108,19 +104,197 @@ export default {
       jsonFields: [],
       expandOnHover: false,
       expandWithDelay: false,
-      reduce: false
+      reduce: false,
+      fileField: {}
     };
   },
+  mounted(){
+    this.initializeRdfForm(this.rdf, this.files);
+    this.fileField = {
+        name: "Files",
+        value: this.files,
+      }
+  },
   methods: {
-    formSubmitted() {
-      // const docstring = DOMPurify.sanitize(marked(docstring));
-    }
+     transformFields(fields) {
+      const typeMapping = {};
+      for (let k in this.components) {
+        typeMapping[this.components[k].name] = k;
+      }
+      // mapping type to component name
+      for (let field of fields) {
+        if (typeMapping[field.type]) {
+          field.is = typeMapping[field.type];
+          delete field.type;
+        }
+      }
+      return fields;
+    },
+    async formSubmitted(result) {
+      const rdfNameMapping = {
+        type: "Type",
+        name: "Name",
+        description: "Description",
+        version: "Version",
+        license: "License",
+        authors: "Authors",
+        // source: "Source",
+        // git_repo: "Git Repository",
+        tags: "Tags"
+        // links: "Links"
+      };
+      const values = result.values;
+      for (let k in rdfNameMapping) {
+        this.rdf[k] = values[rdfNameMapping[k]];
+      }
+      this.zipPackage = new JSZip();
+      // Fix files
+      if (this.zipPackage) {
+        const packageFiles = Object.values(this.zipPackage.files);
+        for (let file of values["Files"]) {
+          if (packageFiles.includes(file)) continue;
+          if (file instanceof Blob) {
+            this.zipPackage.file(file.name, file);
+          } else {
+            console.error("Invalid file type", file);
+          }
+        }
+        // remove files
+        for (let file of packageFiles) {
+          if (!values["Files"].includes(file)) {
+            delete this.zipPackage.files[file.name];
+          }
+        }
+      } else {
+        this.editedFiles = values["Files"];
+      }
+      let rdfFileName = "rdf.yaml";
+
+      this.rdf.type = "dataset";
+      this.rdf.tags = this.rdf.tags || [];
+      this.rdf.config = this.rdf.config || {};
+      this.rdf.config._rdf_file = "./" + rdfFileName;
+      this.rdf.authors = this.rdf.authors.split(",").map(name => {
+        return { name: name.trim() };
+      });
+
+      // TODO: fix attachments.files for the packager
+      const rdf = Object.assign({}, this.rdf);
+      delete rdf._metadata;
+      console.log("RDF: ", rdf);
+      this.rdfYaml = yaml.dump(rdf);
+      const blob = new Blob([this.rdfYaml], {
+        type: "application/yaml"
+      });
+
+      if (this.zipPackage) {
+        delete this.zipPackage.files[rdfFileName];
+        this.zipPackage.file(rdfFileName, blob);
+      } else {
+        const file = new File([blob], rdfFileName);
+        this.editedFiles = this.editedFiles.filter(
+          item => item.name !== rdfFileName
+        );
+        this.editedFiles.push(file);
+      }
+
+      this.similarDeposits = await this.client.getResourceItems({
+        sort: "bestmatch",
+        query: rdf.name
+      });
+      console.log("Similar deposits:", this.similarDeposits);
+      // if there is any similar items, we can try to login first
+      if (this.similarDeposits.length > 0)
+        await this.client.getCredential(true);
+      this.stepIndex = 2;
+    },
+    initializeRdfForm(rdf) {
+      this.rdf = rdf || {};
+      this.rdf.links = this.rdf.links || [];
+      this.rdf.config = this.rdf.config || {};
+      this.jsonFields = this.transformFields([
+        // {
+        //   label: "Files",
+        //   type: "files",
+        //   value: files,
+        //   isRequired: true
+        // },
+        {
+          label: "Name",
+          placeholder: "name",
+          value: this.rdf.name
+        },
+        {
+          label: "Description",
+          placeholder: "description",
+          value: this.rdf.description,
+          help: 'A short description in one sentence' 
+        },
+        {
+          label: "Authors",
+          placeholder: "authors (Full name, separated by comma)",
+          value:
+            this.rdf.authors &&
+            this.rdf.authors.map(author => author.name.split(";")[0]).join(",")
+        },
+        // {
+        //   label: "Source",
+        //   placeholder: "A doi or URL to the source of the item",
+        //   isRequired: false,
+        //   value: this.rdf.version
+        // },
+        {
+          label: "Version",
+          placeholder: "Version in MAJOR.MINOR.PATCH format(e.g. 0.1.0)",
+          isRequired: false,
+          value: this.rdf.version || "0.1.0"
+        },
+        {
+          label: "License",
+          type: "select",
+          placeholder: "Select your license",
+          options: Object.keys(spdxLicenseList).map(opt => {
+            return {
+              text: opt,
+              value: opt,
+              selected: this.rdf.license === opt
+            };
+          })
+        },
+        {
+          label: "Tags",
+          type: "tags",
+          value: this.rdf.tags,
+          placeholder: "Add a tag",
+          options: this.allTags,
+          allow_new: true,
+          icon: "label",
+          isRequired: false
+        },
+        {
+          label: "Detailed Descrription",
+          placeholder: "",
+          type: 'textarea',
+          value: this.rdf.config._details,
+          help: 'Detailed description in markdown format' 
+        },
+        // {
+        //   label: "Links",
+        //   type: "tags",
+        //   value: this.rdf.links,
+        //   placeholder: "Add a link (resource item ID)",
+        //   options: this.resourceItems.map(item => item.id),
+        //   allow_new: true,
+        //   icon: "vector-link",
+        //   isRequired: false
+        // }
+      ]);
+    },
   }
 };
 </script>
 <style>
 .dataset {
-  margin-top: 72px;
   height: 100%;
 }
 .p-1 {
@@ -139,5 +313,9 @@ export default {
 }
 .sidebar-content {
   height: 100% !important;
+}
+.viewer{
+  width: 100%;
+  height: 100%;
 }
 </style>
